Fixed log count and axe damage issues, updated README
Fixed an issue that was causing the plugin to not count the number of logs an axe could break.
Fixed an issue where if an axe should have broke in one hit, it would do strange things with durability due to how Minecraft handles it.
Updated readme to include explanation of config.